requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
